requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pillow==10.2.0
ultralytics==8.1.0
pydantic==2.5.3